+++ jenkins/Jenkinsfile
@@ -51,6 +51,7 @@ pipeline {
 					dir('./services/market') { sh "npm run test:unit" }
 					dir('./services/newsfeed') { sh "npm run test:unit" }
 					dir('./services/export') { sh "npm run test:unit" }
+					dir('./services/gateway') { sh "npm run test:unit" }
 				}
 			}
 		}
